Minor module restructure
