--- bindings/python/iterators/__init__.py
@@ -1,2 +1,0 @@
-from .sentences import SentenceIterator as sentences
-from .words import WordIterator as words
