+++ setup.py
@@ -211,7 +211,7 @@ setup(
     extras_require={"test": ["pytest>=6.0"]},
     license_files=("LICENSE",),
     python_requires=">=3.6",
-    packages=["slimt", "slimt.tests", "slimt.iterators"],
+    packages=["slimt", "slimt.tests"],
     package_dir={"slimt": "bindings/python"},
     install_requires=["pyyaml>=5.1", "appdirs"],
     entry_points={
