+++ slimt/CMakeLists.txt
@@ -99,7 +99,8 @@ foreach(SLIMT_LIB IN LISTS SLIMT_LIBRARIES)
     PUBLIC
       "$<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}>"
       "$<BUILD_INTERFACE:${CMAKE_BINARY_DIR}>"
-      "$<INSTALL_INTERFACE:${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_INCLUDEDIR}>")
+      "$<INSTALL_INTERFACE:${CMAKE_INSTALL_PREFIX}/${CMAKE_INSTALL_INCLUDEDIR}>"
+  )
 
   target_link_options(${SLIMT_LIB} PUBLIC ${SLIMT_LINK_OPTIONS})
   target_compile_options(${SLIMT_LIB} PRIVATE ${SLIMT_COMPILE_OPTIONS})
